modification shell

+++ fuites.h
@@ -11,7 +11,7 @@ typedef struct Chainon {
 } Chainon;
 
 double calculer_pertes(Chainon* c, double vol_entrer);
-void fichier_pertes(char* id, float pertes);
+void fichier_pertes(char* id, double pertes);
